# web/api-docs/swagger.yaml
basePath: /
definitions:
  github_com_cuihe500_vaulthub_internal_database_models.DecryptedSecret:
    properties:
      access_count:
        type: integer
      created_at:
        type: string
      dek_version:
        type: integer
      description:
        type: string
      id:
        type: integer
      last_accessed_at:
        type: string
      metadata:
        $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SecretMetadata'
      plain_data:
        description: 解密后的明文数据
        type: string
      secret_name:
        type: string
      secret_type:
        $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SecretType'
      secret_uuid:
        type: string
      updated_at:
        type: string
      user_uuid:
        type: string
    type: object
  github_com_cuihe500_vaulthub_internal_database_models.SafeEncryptedSecret:
    properties:
      access_count:
        type: integer
      created_at:
        type: string
      dek_version:
        type: integer
      description:
        type: string
      id:
        type: integer
      last_accessed_at:
        type: string
      metadata:
        $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SecretMetadata'
      secret_name:
        type: string
      secret_type:
        $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SecretType'
      secret_uuid:
        type: string
      updated_at:
        type: string
      user_uuid:
        type: string
    type: object
...
  github_com_cuihe500_vaulthub_internal_database_models.SafeUserEncryptionKey:
    properties:
      created_at:
        type: string
      dek_algorithm:
        type: string
      dek_version:
        type: integer
      id:
        type: integer
      kek_algorithm:
        type: string
      last_rotation_at:
        type: string
      rotation_started_at:
        type: string
      rotation_status:
        type: string
      updated_at:
        type: string
      user_uuid:
        type: string
    type: object
...
  github_com_cuihe500_vaulthub_internal_database_models.SecretMetadata:
    properties:
      expires_at:
        description: 过期时间
        type: string
      extra:
        additionalProperties: true
        description: 额外信息
        type: object
      tags:
        description: 标签
        items:
          type: string
        type: array
    type: object
  github_com_cuihe500_vaulthub_internal_database_models.SecretType:
    enum:
    - api_key
    - db_credential
    - certificate
    - ssh_key
    - token
    - password
    - other
    type: string
    x-enum-comments:
      SecretTypeAPIKey: API密钥
      SecretTypeCertificate: 证书
      SecretTypeDBCredential: 数据库凭证
      SecretTypeOther: 其他
      SecretTypePassword: 密码
      SecretTypeSSHKey: SSH密钥
      SecretTypeToken: 令牌
    x-enum-descriptions:
    - API密钥
    - 数据库凭证
    - 证书
    - SSH密钥
    - 令牌
    - 密码
    - 其他
    x-enum-varnames:
    - SecretTypeAPIKey
    - SecretTypeDBCredential
    - SecretTypeCertificate
    - SecretTypeSSHKey
    - SecretTypeToken
    - SecretTypePassword
    - SecretTypeOther
...
  github_com_cuihe500_vaulthub_internal_service.AuditLogDTO:
    properties:
      action_type:
        type: string
      created_at:
        type: string
      details: {}
      error_code:
        type: integer
      error_message:
        type: string
      ip_address:
        type: string
      request_id:
        type: string
      resource_name:
        type: string
      resource_type:
        type: string
      resource_uuid:
        type: string
      status:
        type: string
      user_agent:
        type: string
      user_uuid:
        type: string
      username:
        type: string
      uuid:
        type: string
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.CreateUserEncryptionKeyRequest:
    properties:
      security_pin:
        description: 安全密码，用于保护加密数据（独立于登录密码）
        minLength: 8
        type: string
    required:
    - security_pin
    type: object
  github_com_cuihe500_vaulthub_internal_service.CreateUserEncryptionKeyResponse:
    properties:
      recovery_key:
        description: 恢复密钥（阶段2实现）
        type: string
      user_encryption_key:
        $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUserEncryptionKey'
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.DecryptSecretRequest:
    properties:
      security_pin:
        description: 安全密码，用于解密DEK
        type: string
    required:
    - security_pin
    type: object
  github_com_cuihe500_vaulthub_internal_service.EncryptAndStoreSecretRequest:
    properties:
      description:
        type: string
      metadata:
        $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SecretMetadata'
      plain_data:
        type: string
      secret_name:
        type: string
      secret_type:
        $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SecretType'
      security_pin:
        description: 安全密码，用于解密DEK
        type: string
    required:
    - plain_data
    - secret_name
    - secret_type
    - security_pin
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.ListUserSecretsResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      secrets:
        items:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeEncryptedSecret'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.OperationStatisticsExport:
    properties:
      by_action:
        additionalProperties:
          format: int64
          type: integer
        description: 按操作类型统计
        type: object
      total_operations:
        description: 操作总数
        type: integer
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.SecretStatisticsExport:
    properties:
      by_type:
        additionalProperties:
          format: int64
          type: integer
        description: 按类型统计
        type: object
      total_secrets:
        description: 密钥总数
        type: integer
    type: object
...
  github_com_cuihe500_vaulthub_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      requestId:
        type: string
      timestamp:
        type: integer
    type: object
...
  internal_api_handlers.QueryAuditLogsResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.AuditLogDTO'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@thankseveryone.top
    name: API Support
  description: |-
    VaultHub 是一个密钥管理系统，旨在安全地存储、管理和轮换加密密钥、API 密钥及其他敏感凭证。

    ## 特性
    - 安全的密钥存储和管理
    - 密钥轮换支持
    - 细粒度的访问控制
    - 审计日志

    ## 认证
    大部分接口需要JWT Bearer Token认证。
    1. 调用 /api/v1/auth/register 注册账号
    2. 调用 /api/v1/auth/login 获取 token
    3. 在请求头中添加 Authorization: Bearer {token}

    ## 响应格式
    所有接口返回统一的响应格式，HTTP状态码均为200，错误码在响应体的code字段中：
    ```json
    {
    "code": 0,
    "message": "success",
    "data": {},
    "requestId": "xxx",
    "timestamp": 1762269490888
    }
    ```
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/cuihe500/vaulthub
  title: VaultHub API
  version: "1.0"
paths:
...
  /api/v1/audit/logs:
    get:
      consumes:
      - application/json
      description: 查询审计日志，支持多条件过滤和分页。普通用户只能查询自己的日志，管理员可以查询所有用户的日志。不传分页参数时全量导出（最多10000条）
      parameters:
      - description: 用户UUID（管理员可指定，普通用户自动使用当前用户）
        in: query
        name: user_uuid
        type: string
      - description: 操作类型：CREATE/UPDATE/DELETE/ACCESS/LOGIN/LOGOUT
        in: query
        name: action_type
        type: string
      - description: 资源类型：vault/secret/user/config
        in: query
        name: resource_type
        type: string
      - description: 操作状态：success/failed
        in: query
        name: status
        type: string
      - description: 开始时间（RFC3339格式，如2024-01-01T00:00:00Z）
        in: query
        name: start_time
        type: string
      - description: 结束时间（RFC3339格式）
        in: query
        name: end_time
        type: string
      - description: 页码（可选，不传则全量导出）
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页数量（可选，不传则全量导出）
        in: query
        maximum: 10000
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/internal_api_handlers.QueryAuditLogsResponse'
        "400":
          description: 参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器错误
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 查询审计日志
      tags:
      - 审计
  /api/v1/audit/logs/export:
    get:
      consumes:
      - application/json
      description: 导出各类型加密数据的统计总量。普通用户只能查询自己的统计，管理员可以查询指定用户或全局统计
      parameters:
      - description: 用户UUID（管理员可指定，普通用户自动使用当前用户，管理员不指定则查询全局）
        in: query
        name: user_uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 导出成功
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.SecretStatisticsExport'
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器错误
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 导出密钥类型统计
      tags:
      - 审计
  /api/v1/audit/operations/export:
    get:
      consumes:
      - application/json
      description: 导出指定时间范围内的操作统计数据（按操作类型分组）。普通用户只能查询自己的统计，管理员可以查询指定用户或全局统计
      parameters:
      - description: 用户UUID（管理员可指定，普通用户自动使用当前用户，管理员不指定则查询全局）
        in: query
        name: user_uuid
        type: string
      - description: 开始时间（RFC3339格式，如2024-01-01T00:00:00Z）
        in: query
        name: start_time
        type: string
      - description: 结束时间（RFC3339格式）
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 导出成功
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.OperationStatisticsExport'
        "400":
          description: 参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器错误
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 导出操作统计
      tags:
      - 审计
...
  /api/v1/encryption/keys:
    post:
      consumes:
      - application/json
      description: |-
        为当前用户创建加密密钥（首次使用加密功能时调用）
        警告：返回的恢复密钥仅显示一次，请务必妥善保管
      parameters:
      - description: 创建加密密钥请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.CreateUserEncryptionKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.CreateUserEncryptionKeyResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建用户加密密钥
      tags:
      - 秘密管理
...
  /api/v1/secrets:
    get:
      consumes:
      - application/json
      description: 获取当前用户的秘密列表（不包含加密数据）。不传分页参数时全量导出（最多10000条）
      parameters:
      - description: 秘密类型
        enum:
        - api_key
        - db_credential
        - certificate
        - ssh_key
        - token
        - password
        - other
        in: query
        name: secret_type
        type: string
      - description: 页码（可选，不传则全量导出）
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页数量（可选，不传则全量导出）
        in: query
        maximum: 10000
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.ListUserSecretsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取秘密列表
      tags:
      - 秘密管理
    post:
      consumes:
      - application/json
      description: 加密并存储敏感数据（需要输入密码）
      parameters:
      - description: 创建秘密请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.EncryptAndStoreSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeEncryptedSecret'
              type: object
      security:
      - BearerAuth: []
      summary: 创建加密秘密
      tags:
      - 秘密管理
  /api/v1/secrets/{uuid}:
    delete:
      consumes:
      - application/json
      description: 删除指定的秘密（软删除）
      parameters:
      - description: 秘密UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除秘密
      tags:
      - 秘密管理
  /api/v1/secrets/{uuid}/decrypt:
    post:
      consumes:
      - application/json
      description: 解密并获取秘密的明文数据（需要输入密码）
      parameters:
      - description: 秘密UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: 解密请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.DecryptSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.DecryptedSecret'
              type: object
      security:
      - BearerAuth: []
      summary: 解密秘密
      tags:
      - 秘密管理
...
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer" 后跟一个空格和 JWT token。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"